+++ app/src/main/java/com/emis/emismobile/contact/ContactUsActivity.java
@@ -1,0 +1,16 @@
+package com.emis.emismobile.contact;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.emis.emismobile.R;
+
+public class ContactUsActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_contact_us);
+    }
+}