+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.2.0] - 2020-07-26
+
+### Added
+- Knowledge and Contact Us activities
+- Navigation from the Main activity 
+
+
 ## [0.1.0] - 2020-07-25
 
 ### Configured